Refactor: update README structure, remove Docker watch mode, and enhance middleware organisation
